requirements.txt
pytorch==1.8.1
torchvision==0.9.1
# torchaudio==0.8.1
pytorch_lightning==1.6.0
imageio
imageio-ffmpeg
matplotlib
configargparse
tensorboard>=2.0
tqdm
opencv-python
# spconv-cu111
pycolmap>=0.1.0
scikit-image
scikit-learn
trimesh
h5py
einops
easydict
kornia==0.6.4
numba==0.55.1
omegaconf==2.2.0
yacs==0.1.8
lmdb
inplace_abn
transforms3d
